+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 1.1.0
 
--   Added Security util for url signing.
--   Added [upload](./README.md#upload) method.
+- Added Security util for url signing.
+- Added [upload](./README.md#upload) method.